Prepare U-Boot for gcc-4.x: fix global data pointer initialization
The global data pointer, stored in r29 before relocation to RAM, was
not initialized to a correct value. This happened because the pointer
declaration was done locally in the scope of the board_init_f()
function. What follows is a cite from gcc.info, 5.37, "Variables in
Specified Registers":
    ...
    The compiler's data flow analysis is capable of determining where
    the specified registers contain live values, and where they are
    available for other uses. Stores into local register variables
    may be deleted when they appear to be dead according to dataflow
    analysis. References to local register variables may be deleted
    or moved or simplified.
    ...
Moving the global data declaration to global scope solved the
problem.