Fix possible uninitialized variable compiler warning.
When CONFIG_OF_FLAG_TREE is set, the compiler complains that 'len' in
do_bootm_linux() may be uninitialized.  There is no possibility in the
current code that len will get used uninitialized, but this fix follows
the existing convention of setting both len and data to zero at the same
time.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>