[BUILD] Remove libraries when updating autoconf.mk
Fix library problems caused by conditional compilation.  Using
autoconf.mk to decide which files to compile has caused a problem when
changing configuration from one board to another without clearing out
the library (*.a) files.

It used to be that the linker was always passed the same list of .o
files when building the .a files.  However, that is not longer true
with conditional compilation.  Now, a different board config will have
a different file list passed to the linker.  The problem occurs when
a library has already been built and the board config is changed.

Since the linker will update instead of replace a preexisting library,
then if the file list changes to remove some object files the old
objects will still exist in the library.

The solution is to remove all old library files when autoconf.mk is
made.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>