cmd_bmp: Add support for displaying gzip compressed bmps
The existing code can show information about a gzip compressed BMP
image, but can't actually display it.

Therefore, move the decompression code out of bmp_info() and use it in
bmp_display() as well in order to display a compressed BMP image.

Also, clean things up a bit and fix a memory leak while we're at it.

[hskinnemoen@atmel.com: a bit of refactoring]
Signed-off-by: Haavard Skinnemoen <hskinnemoen@atmel.com>